Fix `codemirror.css` stylesheet address
Project `codemirror` v5 changed all the urls adding `5` infix. So the live demo shows nothing now because styles are not loaded. I consider this fix as a workaround through, because I believe it is recommended to use local copy of `codemirror.css` or use third-party repositories. Another opportunity is to link directly so some github commit of the file.